requirements.txt
streamlit
langchain
langchain-openai
langchain-ollama
langchain-deepseek
mysql-connector-python
python-dotenv
langchain-community
langchain-core
pymysql